feat(security): implemente security in endpoints

+++ cibseven-webclient-core/src/main/java/org/cibseven/webapp/rest/TaskService.java
@@ -1,6 +1,5 @@
 package org.cibseven.webapp.rest;
 
-import java.io.IOException;
 import java.util.Collection;
 import java.util.List;
 import java.util.Locale;
@@ -14,6 +13,7 @@ import org.cibseven.webapp.exception.AccessDeniedException;
 import org.cibseven.webapp.exception.NoObjectFoundException;
 import org.cibseven.webapp.exception.SystemException;
 import org.cibseven.webapp.logger.TaskLogger;
+import org.cibseven.webapp.providers.PermissionConstants;
 import org.cibseven.webapp.providers.SevenProvider;
 import org.cibseven.webapp.rest.model.IdentityLink;
 import org.cibseven.webapp.rest.model.Task;
@@ -39,6 +39,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.core.MediaType;
+
+import static org.cibseven.webapp.auth.SevenAuthorizationUtils.*;
 
 @ApiResponses({
 	@ApiResponse(responseCode = "500", description = "An unexpected system error occured"),
@@ -74,12 +76,14 @@ public class TaskService extends BaseService implements InitializingBean {
 			@Parameter(description = "Task definition key in") @RequestParam Optional<String> taskDefinitionKeyIn,
 			Locale loc, HttpServletRequest rq) {
 		CIBUser user = checkAuthorization(rq, true, false);
+		checkPermission(user, TASK, PermissionConstants.READ_ALL);
 		return sevenProvider.findTasksCount(name, nameLike, taskDefinitionKey, taskDefinitionKeyIn, user);
 	}
 	 
 	//Not used
 	@RequestMapping(value = "/task/by-process-instance/{processInstanceId}", method = RequestMethod.GET)
 	public Collection<Task> findTasksByProcessInstance(@PathVariable String processInstanceId, Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.READ_ALL);
 		return bpmProvider.findTasksByProcessInstance(processInstanceId, user);
 	}
 
@@ -103,6 +107,7 @@ public class TaskService extends BaseService implements InitializingBean {
 	public Task findTaskById(
 			@Parameter(description = "Task Id") @PathVariable String taskId,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.READ_ALL);
 		return bpmProvider.findTaskById(taskId, user);
 	}
 	
@@ -118,6 +123,7 @@ public class TaskService extends BaseService implements InitializingBean {
 			@Parameter(description = "Index of the first result to return") @RequestParam Integer firstResult,
 			@Parameter(description = "Maximum number of results to return") @RequestParam Integer maxResults,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.READ_ALL);
 		return bpmProvider.findTasksByFilter(filters, filterId, user, firstResult, maxResults);
 	}
 	
@@ -131,6 +137,7 @@ public class TaskService extends BaseService implements InitializingBean {
 			@Parameter(description = "Filter Id") @PathVariable String filterId,
 			@RequestBody TaskFiltering filters,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.READ_ALL);
 		return bpmProvider.findTasksCountByFilter(filterId, user, filters);
 	}
 	
@@ -143,6 +150,7 @@ public class TaskService extends BaseService implements InitializingBean {
 	public void submit(
 			@Parameter(description = "Task Id") @PathVariable String taskId,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.UPDATE_ALL);
 		bpmProvider.submit(taskId, user);
 	}
 	
@@ -154,6 +162,7 @@ public class TaskService extends BaseService implements InitializingBean {
 	public Object formReference(
 			@Parameter(description = "Task Id") @PathVariable String taskId,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.READ_ALL);
 		return bpmProvider.formReference(taskId, user);
 	}
 	
@@ -166,6 +175,7 @@ public class TaskService extends BaseService implements InitializingBean {
 			@Parameter(description = "Task Id") @PathVariable String taskId,
 			@Parameter(description = "User to be set as assignee") @PathVariable String userId,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.UPDATE_ALL);
 		bpmProvider.setAssignee(taskId, userId, user);
 	}
 	
@@ -177,6 +187,7 @@ public class TaskService extends BaseService implements InitializingBean {
 	public void update(
 			@RequestBody Task task,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.UPDATE_ALL);
 		bpmProvider.update(task, user);
 	}
 
@@ -192,6 +203,7 @@ public class TaskService extends BaseService implements InitializingBean {
 			@RequestBody Map<String, Object> data,
 			Locale loc, HttpServletRequest rq) {
 		CIBUser user = checkAuthorization(rq, true, false);
+		checkPermission(user, TASK, PermissionConstants.READ_ALL);
 		return sevenProvider.findTasksPost(data, user);
 	}
 
@@ -204,6 +216,7 @@ public class TaskService extends BaseService implements InitializingBean {
 			@Parameter(description = "Task Id") @PathVariable String taskId,
 			@Parameter(description = "Type of links to include e.g. 'candidate'") @RequestParam Optional<String> type,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.READ_ALL);
 		return bpmProvider.findIdentityLink(taskId, type, user);
 	}
 	
@@ -215,6 +228,7 @@ public class TaskService extends BaseService implements InitializingBean {
 			@Parameter(description = "Task Id") @PathVariable String taskId,
 			@RequestBody Map<String, Object> data,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.UPDATE_ALL);
 		bpmProvider.createIdentityLink(taskId, data, user);
 	}
 	
@@ -226,6 +240,7 @@ public class TaskService extends BaseService implements InitializingBean {
 			@Parameter(description = "Task Id") @PathVariable String taskId,
 			@RequestBody Map<String, Object> data,
 			Locale loc, CIBUser user) {
+		checkPermission(user, TASK, PermissionConstants.DELETE_ALL);
 		bpmProvider.deleteIdentityLink(taskId, data, user);
 	}
 	
@@ -238,6 +253,7 @@ public class TaskService extends BaseService implements InitializingBean {
 			@Parameter(description = "Activity instance Id") @PathVariable String activityInstanceId,
 			Locale loc, CIBUser user) {
 		checkCockpitRights(user);
+        checkPermission(user, PROCESS_INSTANCE, PermissionConstants.READ_ALL);
 		return bpmProvider.fetchActivityVariables(activityInstanceId, user);
 	}
 	
@@ -247,6 +263,7 @@ public class TaskService extends BaseService implements InitializingBean {
 	@ApiResponse(responseCode = "404", description= "Variable name not found")
 	@RequestMapping(value = "/task/{processInstanceId}/variable/download/{variableName}", method = RequestMethod.GET)
 	public ResponseEntity<byte[]> downloadFiles(@PathVariable String processInstanceId, @PathVariable String variableName, CIBUser user) {
+        checkPermission(user, PROCESS_INSTANCE, PermissionConstants.READ_ALL);
 		return bpmProvider.fetchProcessInstanceVariableData(processInstanceId, variableName, user);
 	}
 	
@@ -259,6 +276,7 @@ public class TaskService extends BaseService implements InitializingBean {
 	    TaskLogger logger = new TaskLogger(processDefinitionId, processInstanceId, taskName, taskId);
 	    try {
 	      CIBUser userAuth = (CIBUser) baseUserProvider.authenticateUser(rq);
+          checkPermission(userAuth, PROCESS_INSTANCE, PermissionConstants.UPDATE_ALL);
 	      logger.info("[INFO] Submit variables in task with name=" + taskName + " and ID=" + taskId + " (" + getClass().getSimpleName() + ")");
 	      Task task = bpmProvider.findTaskById(taskId, userAuth);
 	      if (!task.getAssignee().equals(userAuth.getUserID())) {
@@ -291,6 +309,7 @@ public class TaskService extends BaseService implements InitializingBean {
 	  public Variable fetchVariable(@PathVariable String taskId, @PathVariable String variableName, 
 	      @RequestParam Optional<Boolean> deserialize, HttpServletRequest rq) {
 	    CIBUser userAuth = (CIBUser) baseUserProvider.authenticateUser(rq);
+        checkPermission(userAuth, PROCESS_INSTANCE, PermissionConstants.READ_ALL);
 	    return bpmProvider.fetchVariable(taskId, variableName, deserialize, userAuth);
 	  }
 	  
@@ -298,6 +317,7 @@ public class TaskService extends BaseService implements InitializingBean {
 	  public byte[] fetchVariableFileData(@PathVariable String taskId, @PathVariable String variableName, 
 	      @RequestParam Optional<Boolean> deserialize, HttpServletRequest rq) {
 	    CIBUser userAuth = (CIBUser) baseUserProvider.authenticateUser(rq);
+        checkPermission(userAuth, TASK, PermissionConstants.READ_ALL);
 	    NamedByteArrayDataSource res = bpmProvider.fetchVariableFileData(taskId, variableName, userAuth);
 	    return res.getContent();
 	  }
@@ -306,18 +326,21 @@ public class TaskService extends BaseService implements InitializingBean {
 	  public Map<String, Variable> fetchVariables(@PathVariable String taskId, 
 	      @RequestParam Optional<Boolean> deserialize, 
 	      @RequestParam Optional<String> locale, CIBUser user) throws Exception {
+        checkPermission(user, TASK, PermissionConstants.READ_ALL);
 	    return bpmProvider.fetchFormVariables(taskId, deserialize.orElse(false), user);
 	  }
 
 	  @RequestMapping(value = "/task/{taskId}/variable/{variableName}", method = RequestMethod.DELETE)
 	  public void deleteVariable(@PathVariable String taskId, @PathVariable String variableName, HttpServletRequest rq) {
 	    CIBUser userAuth = (CIBUser) baseUserProvider.authenticateUser(rq);
+        checkPermission(userAuth, TASK, PermissionConstants.DELETE_ALL);
 	    bpmProvider.deleteVariable(taskId, variableName, userAuth);
 	  }
 	  
 	  @RequestMapping(value = "/task/{taskId}/bpmnError", method = RequestMethod.POST)
 	  public void handleBpmnError(@PathVariable String taskId, @RequestBody Map<String, Object> data, 
 	      @RequestParam Optional<String> locale, CIBUser user) throws Exception {
+		checkPermission(user, TASK, PermissionConstants.UPDATE_ALL);
 	    bpmProvider.handleBpmnError(taskId, data, user);
 	  }
